aded set state code
